refactor: move types to common and json protocol
