+++ minecat/cogs/client/connect.py
@@ -9,13 +9,15 @@ from typing import TYPE_CHECKING
 
 from botbase import CogBase
 from orjson import loads
-from websockets.client import connect, WebSocketClientProtocol
+from common import JsonWebSocketClient
+from mineager import Opcode
+from websockets.client import WebSocketClientProtocol, connect
 from websockets.exceptions import ConnectionClosed
 
 if TYPE_CHECKING:
     from minecat.__main__ import Minecat
-    from ..websocket import JsonType
 
+    from ..websocket import JsonType
 
 log = getLogger(__name__)
 
@@ -27,7 +29,11 @@ class Connect(CogBase["Minecat"]):
         bot.loop.create_task(self.connect())
 
     async def connect(self) -> None:
-        async for ws in connect("ws://manager:6420"):
+        async for ws in connect(
+            "ws://manager:6420",
+            create_protocol=JsonWebSocketClient,  # type: ignore  # doesnt know how to type
+        ):
+            await ws.send(f'{{"o": {Opcode.LOGIN.value}, "d": {self.bot.cluster}}}')
             try:
                 async for message in ws:
                     await self.handler(ws, loads(message))
@@ -36,6 +42,10 @@ class Connect(CogBase["Minecat"]):
 
     async def handler(self, ws: WebSocketClientProtocol, message: JsonType) -> None:
         log.debug("Received message from manager: %s", message)
+        opcode = Opcode(message["o"])
+
+        if opcode == Opcode.RESTART:
+            await self.bot.close()
 
 
 def setup(bot: Minecat):
