+++ minecat/cogs/websocket/__init__.py
@@ -8,4 +8,3 @@ from ._errors import *
 from ._manager import *
 from ._opcode import *
 from ._server import *
-from ._types import *
